Rename: 게시글 관련 컴포넌트명 및 폴더구조 변경
